trying macos again

+++ recipe/build.sh
@@ -19,45 +19,61 @@ export GMP_LIB_DIRS=$PREFIX/lib
 # Install shim scripts to ensure that certain flags are always passed to the compiler/linker
 #
 
-echo "#!/bin/bash" > $CC-shim
-echo "set -e -o pipefail -x " >> $CC-shim
-echo "$CC -I$PREFIX/include -L$PREFIX/lib -pthread -fPIC \"\$@\"" >> $CC-shim
-chmod u+x $CC-shim
-export CC=$CC-shim
+if [ -n "${CC+set}" ]; then
+    echo "#!/bin/bash" > $CC-shim
+    echo "set -e -o pipefail -x " >> $CC-shim
+    echo "$CC -I$PREFIX/include -L$PREFIX/lib -pthread -fPIC \"\$@\"" >> $CC-shim
+    chmod u+x $CC-shim
+    export CC=$CC-shim
+fi
 
-echo "#!/bin/bash" > $CXX-shim
-echo "set -e -o pipefail -x " >> $CXX-shim
-echo "$CXX -I$PREFIX/include -L$PREFIX/lib -pthread -fPIC \"\$@\"" >> $CXX-shim
-chmod u+x $CXX-shim
-export CXX=$CXX-shim
 
-echo "#!/bin/bash" > $GCC-shim
-echo "set -e -o pipefail -x " >> $GCC-shim
-echo "$GCC -I$PREFIX/include -L$PREFIX/lib -pthread -fPIC \"\$@\"" >> $GCC-shim
-chmod u+x $GCC-shim
-export GCC=$GCC-shim
+if [ -n "${CXX+set}" ]; then
+    echo "#!/bin/bash" > $CXX-shim
+    echo "set -e -o pipefail -x " >> $CXX-shim
+    echo "$CXX -I$PREFIX/include -L$PREFIX/lib -pthread -fPIC \"\$@\"" >> $CXX-shim
+    chmod u+x $CXX-shim
+    export CXX=$CXX-shim
+fi
 
-echo "#!/bin/bash" > $GXX-shim
-echo "set -e -o pipefail -x " >> $GXX-shim
-echo "$GXX -I$PREFIX/include -L$PREFIX/lib -pthread -fPIC \"\$@\"" >> $GXX-shim
-chmod u+x $GXX-shim
-export GXX=$GXX-shim
+if [ -n "${GCC+set}" ]; then
+    echo "#!/bin/bash" > $GCC-shim
+    echo "set -e -o pipefail -x " >> $GCC-shim
+    echo "$GCC -I$PREFIX/include -L$PREFIX/lib -pthread -fPIC \"\$@\"" >> $GCC-shim
+    chmod u+x $GCC-shim
+    export GCC=$GCC-shim
+fi
 
-echo "#!/bin/bash" > $LD-shim
-echo "set -e -o pipefail -x " >> $LD-shim
-echo "$LD -L$PREFIX/lib \"\$@\"" >> $LD-shim
-chmod u+x $LD-shim
-export LD=$LD-shim
 
-echo "#!/bin/bash" > ${LD}.gold
-echo "set -e -o pipefail -x " >> ${LD}.gold
-echo "$LD_GOLD -L$PREFIX/lib \"\$@\"" >> ${LD}.gold
-chmod u+x ${LD}.gold
-export LD_GOLD=${LD}.gold
+if [ -n "${GXX+set}" ]; then
+    echo "#!/bin/bash" > $GXX-shim
+    echo "set -e -o pipefail -x " >> $GXX-shim
+    echo "$GXX -I$PREFIX/include -L$PREFIX/lib -pthread -fPIC \"\$@\"" >> $GXX-shim
+    chmod u+x $GXX-shim
+    export GXX=$GXX-shim
+fi
+
+if [ -n "${LD+set}" ]; then
+    echo "#!/bin/bash" > $LD-shim
+    echo "set -e -o pipefail -x " >> $LD-shim
+    echo "$LD -L$PREFIX/lib \"\$@\"" >> $LD-shim
+    chmod u+x $LD-shim
+    export LD=$LD-shim
+fi
+
+if [ -n "${LD_GOLD+set}" ]; then
+    echo "#!/bin/bash" > ${LD}.gold
+    echo "set -e -o pipefail -x " >> ${LD}.gold
+    echo "$LD_GOLD -L$PREFIX/lib \"\$@\"" >> ${LD}.gold
+    chmod u+x ${LD}.gold
+    export LD_GOLD=${LD}.gold
+fi
 
 HOST_LIBPTHREAD="${BUILD_PREFIX}/${HOST}/sysroot/usr/lib/libpthread.so"
-rm ${HOST_LIBPTHREAD}
-ln -s /lib64/libpthread.so.0 ${HOST_LIBPTHREAD}
+if [ -f "$HOST_LIBPTHREAD" ]; then
+    rm ${HOST_LIBPTHREAD}
+    ln -s /lib64/libpthread.so.0 ${HOST_LIBPTHREAD}
+fi
 
 #######################################################################################################
 # Build git-annex
@@ -78,5 +94,5 @@ mkdir -p $STACK_ROOT
 cp ${RECIPE_DIR}/git_annex_custom_snapshot.yaml .
 stack setup --resolver git_annex_custom_snapshot.yaml
 stack update --resolver git_annex_custom_snapshot.yaml
-stack install --resolver git_annex_custom_snapshot.yaml --extra-include-dirs ${PREFIX}/include --extra-lib-dirs ${PREFIX}/lib --ghc-options " -optc-I${PREFIX}/include -optl-L${PREFIX}/lib " --local-bin-path ${PREFIX}/bin --flag git-annex:magicmime --flag git-annex:dbus
+stack install --resolver git_annex_custom_snapshot.yaml --extra-include-dirs ${PREFIX}/include --extra-lib-dirs ${PREFIX}/lib --ghc-options " -optc-I${PREFIX}/include -optl-L${PREFIX}/lib " --local-bin-path ${PREFIX}/bin
 ln -s ${PREFIX}/bin/git-annex ${PREFIX}/bin/git-annex-shell
